Feign client with hystrix

+++ ProductManagement/src/main/java/com/cts/repository/ProductManagementRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.cts.entity.Product;
 
 @Repository
-public interface ProductManagementRepository extends CrudRepository<Product, Integer> {
+public interface ProductManagementRepository extends CrudRepository<Product, Long> {
 
 	List<Product> findAllByProductName(String productName);
 	List<Product> findAllByVendorVendorId(Long id);
